fix:헤더 로그인기능 변경
