+++ src/components/Common/Header.jsx
@@ -1,4 +1,6 @@
 import { styled } from 'styled-components';
+import React, { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 const HeaderContainer = styled.div`
   width: 100vw;
@@ -30,10 +32,39 @@ const LoginButton = styled.button`
 `;
 
 function Header() {
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  useEffect(() => {
+    const token = localStorage.getItem('token');
+    if (token) {
+      setIsLoggedIn(true);
+    } else {
+      setIsLoggedIn(false);
+    }
+  }, []);
+  const navigate = useNavigate();
+
+  const handleLoginClick = () => {
+    navigate('/login');
+  };
+
+  const handleLogoutClick = () => {
+    localStorage.removeItem('token');
+    setIsLoggedIn(false);
+    alert('로그아웃 했습니다.');
+    navigate('/login');
+  };
   return (
     <HeaderContainer>
       <h1>Croup</h1>
-      <LoginButton type="button">LogIn</LoginButton>
+      {isLoggedIn ? (
+        <LoginButton type="button" onClick={handleLogoutClick}>
+          Logout
+        </LoginButton>
+      ) : (
+        <LoginButton type="button" onClick={handleLoginClick}>
+          LogIn
+        </LoginButton>
+      )}
     </HeaderContainer>
   );
 }
