+++ src/components/Home/HeaderComponent.jsx
@@ -1,28 +1,28 @@
-import React, { useEffect, useState } from 'react';
+// import React, { useEffect, useState } from 'react';
 import { styled } from 'styled-components';
-import { useNavigate } from 'react-router-dom';
+// import { useNavigate } from 'react-router-dom';
 
 const HeaderContainer = styled.div`
   width: 1300px;
   display: flex;
   flex-direction: column;
 `;
-const HighWrapper = styled.div`
-  display: flex;
-  justify-content: right;
-  margin: 10px 10px;
-  > button {
-    width: 85px;
-    height: 30px;
-    border-radius: 30px;
-    color: white;
-    background-color: #04314d;
-    font-weight: bold;
-    border: 1px solid white;
-    margin-right: 40px;
-    cursor: pointer;
-  }
-`;
+// const HighWrapper = styled.div`
+//   display: flex;
+//   justify-content: right;
+//   margin: 10px 10px;
+//   > button {
+//     width: 85px;
+//     height: 30px;
+//     border-radius: 30px;
+//     color: white;
+//     background-color: #04314d;
+//     font-weight: bold;
+//     border: 1px solid white;
+//     margin-right: 40px;
+//     cursor: pointer;
+//   }
+// `;
 const MiddleWrapper = styled.div`
   pointer-events: none;
   border-radius: 30px;
@@ -61,41 +61,30 @@ const LowWrapper = styled.div`
 `;
 
 function HeaderComponent() {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  // const [isLoggedIn, setIsLoggedIn] = useState(false);
 
-  useEffect(() => {
-    const token = localStorage.getItem('token');
-    if (token) {
-      setIsLoggedIn(true);
-    } else {
-      setIsLoggedIn(false);
-    }
-  }, []);
-  const navigate = useNavigate();
+  // useEffect(() => {
+  //   const token = localStorage.getItem('token');
+  //   if (token) {
+  //     setIsLoggedIn(true);
+  //   } else {
+  //     setIsLoggedIn(false);
+  //   }
+  // }, []);
+  // const navigate = useNavigate();
 
-  const handleLoginClick = () => {
-    navigate('/login');
-  };
+  // const handleLoginClick = () => {
+  //   navigate('/login');
+  // };
 
-  const handleLogoutClick = () => {
-    localStorage.removeItem('token');
-    setIsLoggedIn(false);
-    alert('로그아웃 했습니다.');
-    navigate('/login');
-  };
+  // const handleLogoutClick = () => {
+  //   localStorage.removeItem('token');
+  //   setIsLoggedIn(false);
+  //   alert('로그아웃 했습니다.');
+  //   navigate('/login');
+  // };
   return (
     <HeaderContainer>
-      <HighWrapper>
-        {isLoggedIn ? (
-          <button type="button" onClick={handleLogoutClick}>
-            Logout
-          </button>
-        ) : (
-          <button type="button" onClick={handleLoginClick}>
-            Login
-          </button>
-        )}
-      </HighWrapper>
       <MiddleWrapper>
         <div>
           <p>
