+++ src/components/Home/RoomComponent.jsx
@@ -91,7 +91,7 @@ function RoomComponent({ roomTitle, roomDescription, roomThumbnail, roomId }) {
       });
     } else {
       console.error('토큰이 없습니다.');
-      navigate('/');
+      navigate('/login');
     }
   };
